third_party/WebKit/LayoutTests/virtual/off-main-thread-fetch/http/tests/misc/resource-timing-sizes-cache-worker-expected.txt

This is a testharness.js-based test.
FAIL PerformanceResourceTiming sizes caching test assert_equals: Wrong number of entries expected 3 but got 0
Harness: the test ran to completion.
